+++ src/Services/Helios/DGFiPConnection/DGFiPConnectorOnFTP.php
@@ -97,22 +97,18 @@ class DGFiPConnectorOnFTP implements DGFiPConnector
     }
 
     /**
-     * Télécharge un fichier
-     * @param string $tmp_file chemin du fichier local
-     * @param string $file chemin du fichier distant
-     * @return bool
+     * @throws \S2low\Services\Helios\DGFiPConnection\FTPFileRetrieveException
      * @throws \Exception
      */
-    public function retrieveFile(string $tmp_file, string $file): bool
+    public function retrieveFile(string $localFilePath, string $remoteFilePath): void
     {
-        $ftp_get_result = $this->getConnection()->get($tmp_file, "$file");
+        $ftp_get_result = $this->getConnection()->get($localFilePath, "$remoteFilePath");
         if (!$ftp_get_result) {
             $errorMessage = var_export(error_get_last(), true);
-            throw new Exception(
-                "Impossible de transférer le fichier distant $file vers $tmp_file : " . $errorMessage
+            throw new FTPFileRetrieveException(
+                "Impossible de transférer le fichier distant $remoteFilePath vers $localFilePath : " . $errorMessage
             );
         }
-        return true;
     }
 
     /**