+++ src/Services/Helios/DGFiPConnection/FTPFileRetrieveException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace S2low\Services\Helios\DGFiPConnection;
+
+class FTPFileRetrieveException extends \Exception
+{
+}