+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Permettre de convertir les login/mot de passe de l'identification HTTP de l'ISO-8859-1 vers l'UTF-8 en spécifiant CONVERT_API_LOGINS_FROM_ISO #1071
 - Permettre de convertir les login/mot de passe contenus dans le nonce de l'ISO-8859-1 vers l'UTF-8 en spécifiant CONVERT_API_LOGINS_FROM_ISO #1071
+- Continuer la récupération des PES depuis le serveur DGFiP après un erreur de téléchargement #1183
 
 ## 5.0.37 - 2024-04-03